Fix incorrect locknut orientation
Since we assume that the screw originates from where the nut is, and
runs up towards the head, the locknut should be oriented in the opposite
direction.

This also simplifies the `nutcatch_parallel` case (a parallel nutcatch
for a locknut turns out to look just like regular one), while slightly
complicating the sidecut case.

Turn `nut` into a more private-sounding `_nut`. We don't intend to
expose this function as a public API for the time being.